requirements.txt
bcrypt==5.0.0
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
Markdown==3.10.1
Pygments==2.19.2
python-dotenv==1.2.1
python-frontmatter==1.1.0
PyYAML==6.0.3
requests==2.32.5
urllib3==2.6.3
